fix comments, format, unfocus test
