+++ projects/gateway/pkg/translator/translator_test.go
@@ -1,4 +1,4 @@
-package translator
+package translator_test
 
 import (
 	"context"
@@ -10,6 +10,7 @@ import (
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 	v2 "github.com/solo-io/gloo/projects/gateway/pkg/api/v2"
+	. "github.com/solo-io/gloo/projects/gateway/pkg/translator"
 	"github.com/solo-io/gloo/projects/gloo/pkg/api/v1/plugins/tcp"
 
 	v1 "github.com/solo-io/gloo/projects/gateway/pkg/api/v1"
@@ -768,103 +769,6 @@ var _ = Describe("Translator", func() {
 				})
 			})
 
-			Context("bad config on a delegate route", func() {
-				It("returns the correct error", func() {
-					type routeErr struct {
-						route       *v1.Route
-						expectedErr error
-					}
-					for _, badRoute := range []routeErr{
-						{
-							route: &v1.Route{
-								Matcher: &gloov1.Matcher{
-									PathSpecifier: &gloov1.Matcher_Regex{
-										Regex: "/any",
-									},
-								},
-								Action: &v1.Route_DelegateAction{
-									DelegateAction: &core.ResourceRef{
-										Name:      "any",
-										Namespace: ns,
-									},
-								},
-							},
-							expectedErr: missingPrefixErr,
-						},
-						{
-							route: &v1.Route{
-								Matcher: &gloov1.Matcher{
-									PathSpecifier: &gloov1.Matcher_Exact{
-										Exact: "/any",
-									},
-								},
-								Action: &v1.Route_DelegateAction{
-									DelegateAction: &core.ResourceRef{
-										Name:      "any",
-										Namespace: ns,
-									},
-								},
-							},
-							expectedErr: missingPrefixErr,
-						},
-						{
-							route: &v1.Route{
-								Matcher: &gloov1.Matcher{
-									PathSpecifier: &gloov1.Matcher_Prefix{
-										Prefix: "/any",
-									},
-									Headers: []*gloov1.HeaderMatcher{{}},
-								},
-								Action: &v1.Route_DelegateAction{
-									DelegateAction: &core.ResourceRef{
-										Name:      "any",
-										Namespace: ns,
-									},
-								},
-							},
-							expectedErr: hasHeaderMatcherErr,
-						},
-						{
-							route: &v1.Route{
-								Matcher: &gloov1.Matcher{
-									PathSpecifier: &gloov1.Matcher_Prefix{
-										Prefix: "/any",
-									},
-									Methods: []string{"any"},
-								},
-								Action: &v1.Route_DelegateAction{
-									DelegateAction: &core.ResourceRef{
-										Name:      "any",
-										Namespace: ns,
-									},
-								},
-							},
-							expectedErr: hasMethodMatcherErr,
-						},
-						{
-							route: &v1.Route{
-								Matcher: &gloov1.Matcher{
-									PathSpecifier: &gloov1.Matcher_Prefix{
-										Prefix: "/any",
-									},
-									QueryParameters: []*gloov1.QueryParameterMatcher{{}},
-								},
-								Action: &v1.Route_DelegateAction{
-									DelegateAction: &core.ResourceRef{
-										Name:      "any",
-										Namespace: ns,
-									},
-								},
-							},
-							expectedErr: hasQueryMatcherErr,
-						},
-					} {
-						rv := &routeVisitor{}
-						_, err := rv.convertDelegateAction(badRoute.route)
-						Expect(err).To(Equal(badRoute.expectedErr))
-					}
-				})
-			})
 		})
 	})
